Fix Seat 5.0

+++ src/database/migrations/2017_02_01_093144_create_calendar_tables.php
@@ -16,7 +16,7 @@ class CreateCalendarTables extends Migration
         Schema::create('calendar_operations', function (Blueprint $table): void {
             $table->increments('id');
 
-            $table->bigInteger('user_id');
+            $table->unsignedInteger('user_id');
 
             $table->string('title');
             $table->timestamp('start_at')->nullable();
@@ -45,7 +45,7 @@ class CreateCalendarTables extends Migration
 
             $table->integer('operation_id')->unsigned();
 
-            $table->bigInteger('user_id');
+            $table->unsignedInteger('user_id');
 
             $table->bigInteger('character_id');
             $table->enum('status', ['yes', 'no', 'maybe']);
